changed item price from integer to float

+++ db/migrate/20190503203422_create_items.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 class CreateItems < ActiveRecord::Migration[5.2]
   def change
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.integer :price
+      t.float :price
       t.string :category
       t.string :imgURL
